Import remark as dependency
Previously `remark-parse` and `remark-stringify` were loaded as plugins.
Now remark is loaded as a regular dependency. The difference is that the
language server now uses its own bundled remark version instead of
relying on these plugins being available in the workspace.

This means the language server now doesn’t show an error if
`remark-parse` or `remark-stringify` isn’t available in the workspace.
This is mostly notable if the language server is started automatically
by a language client in a project which doesn’t use remark.